requirements.txt
playwright>=1.40.0
PyYAML>=6.0
urllib3~=1.26.0
typing-extensions>=4.5.0
rich
questionary
readability-lxml
beautifulsoup4
lxml
requests
markdownify
pandas
html5lib